Add files via upload
this zip contains excel files of collected betting data and its corresponding martingale system programmed using c.  i used it for betting research and bankroll management. i am planning to collect more data and build an AI algorithm to predict the sports outcome for profit. you're welcome to come and work on it with me.